--- a/Cargo.toml
+++ b/Cargo.toml
@@ -13,8 +13,8 @@ keywords = ["hashmap", "u64", "intmap"]
 serde = { version = "1.0", optional = true, default-features = false }
 
 [dev-dependencies]
-rand = "0.3"
-ordermap = "0.2.7"
+rand = "0.8.5"
+indexmap = "1.8.2"
 
 [package.metadata.docs.rs]
 features = ["serde"]