--- a/benches/basic_bench.rs
+++ b/benches/basic_bench.rs
@@ -13,6 +13,7 @@ use std::collections::HashMap;
 #[cfg(test)]
 mod tests {
     use super::*;
+    use intmap::Entry;
     use test::Bencher;
 
     const VEC_COUNT: usize = 1000;
@@ -96,7 +97,25 @@ mod tests {
         b.iter(|| {
             map.clear();
             for s in data.iter() {
-                test::black_box({ map.insert(*s, s) });
+                test::black_box(map.insert(*s, s));
+            }
+        });
+    }
+
+    #[bench]
+    fn u64_insert_intmap_entry(b: &mut Bencher) {
+        let data = get_random_range(VEC_COUNT);
+
+        let mut map = IntMap::with_capacity(data.len());
+
+        b.iter(|| {
+            map.clear();
+
+            for s in data.iter() {
+                test::black_box(match map.entry(*s) {
+                    Entry::Occupied(_) => panic!("unexpected while insert, i = {}", s),
+                    Entry::Vacant(entry) => entry.insert(s),
+                });
             }
         });
     }