@JesperAxelsson 2022-02-23
First of all, thanks for taking your time to implement this! 

I'm still pretty inexperienced with Rust and especially unsafe code. So yeah, the unsafe stuff does make me a bit uncomfortable ^^

My plan is to take a closer look this weekend or next week as time allows and compare to the official HashMap implementation. And to better understand the Entry api as well.