@JesperAxelsson 2022-06-08
Sorry for the delay in responding. 

So some thoughts:

In the `Entry` implementation I think it might be better to just use the whole `IntMap` as argument to `new`, you already send most of the data anyway. This would also mean you can use the functions on the `IntMap` rather then having them as free functions.

In `Insert`, `get_mut` and `remove` I would still prefer the old implementation. While the speed is still the same you are putting more object on the stack then necessary.

Now on to the bugbear: unsafe! From my benchmarks I could not tell a difference between the unsafe and safe version. My suggestion would be to create a safe version that is enabled by default and hide the unsafe version behind a feature flag. This would also alleviate those who have string feelings against unsafe code.

If you have lost interest in the project (which I would understand) I might implement these changes myself when I get the time.